backend/requirements.txt
Django>=3.0,<4.0
djangorestframework
django-cors-headers
pygame
channels
channels-redis
daphne
python-socketio
eventlet
socketio
psycopg2